Proposition about structure of MUI components
