+++ src/mui/index.js
@@ -1,0 +1,4 @@
+import Button from './Button'
+import Radio from './Radio'
+
+export { Button, Radio }
