+++ src/utils/jss.js
@@ -1,0 +1,1 @@
+export { withStyles } from '@material-ui/core/styles'
